Feat: [EVER-129] [EVER-129] 인증 기반 쿠폰 사용 API 구현

+++ src/main/java/com/team4ever/backend/domain/coupon/entity/UserCoupon.java
@@ -5,8 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Table(name = "user_coupons")
 @Getter
@@ -19,7 +17,7 @@ public class UserCoupon {
     private Integer id;
 
     @Column(columnDefinition = "INT UNSIGNED COMMENT '유저 FK'")
-    private Integer userId;
+    private Long userId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
@@ -31,7 +29,7 @@ public class UserCoupon {
 
     private Boolean isUsed = false;
 
-    public static UserCoupon of(Integer userId, Coupon coupon) {
+    public static UserCoupon of(Long userId, Coupon coupon) {
         UserCoupon uc = new UserCoupon();
         uc.userId = userId;
         uc.coupon = coupon;
